--- a/encoding.go
+++ b/encoding.go
@@ -68,6 +68,12 @@ func readKeyFromEntry(data []byte) string {
 	return bytesToString(dst)
 }
 
+func compareKeyFromEntry(data []byte, key string) bool {
+	length := binary.LittleEndian.Uint16(data[timestampSizeInBytes+hashSizeInBytes:])
+
+	return bytesToString(data[headersSizeInBytes:headersSizeInBytes+length]) == key
+}
+
 func readHashFromEntry(data []byte) uint64 {
 	return binary.LittleEndian.Uint64(data[timestampSizeInBytes:])
 }