--- a/shard.go
+++ b/shard.go
@@ -104,10 +104,10 @@ func (s *cacheShard) getValidWrapEntry(key string, hashedKey uint64) ([]byte, er
 		return nil, err
 	}
 
-	if entryKey := readKeyFromEntry(wrappedEntry); key != entryKey {
+	if compareKeyFromEntry(wrappedEntry, key) {
 		s.collision()
 		if s.isVerbose {
-			s.logger.Printf("Collision detected. Both %q and %q have the same hash %x", key, entryKey, hashedKey)
+			s.logger.Printf("Collision detected. Both %q and %q have the same hash %x", key, readKeyFromEntry(wrappedEntry), hashedKey)
 		}
 
 		return nil, ErrEntryNotFound