The append function is slow compared to the get and set method. The reason for this is that unnecessary copies of the data are done in the process. Namely, the `get` copy the data and the list `append` a second time. This leads to high allocation numbers and slow runtime overall.
In this PR the number of copies in the process is minimized without touching the `BytesQueue` logic. This reduces the allocs and halves the runtime of append without changing on the interface.
Some benchmarks result:
```
name                          old time/op    new time/op    delta
AppendToCache/1-shards-12       8.37µs ±26%    4.06µs ±23%  -51.48%  (p=0.000 n=99+98)
AppendToCache/512-shards-12     3.23µs ± 6%    1.87µs ± 3%  -42.00%  (p=0.000 n=96+92)
AppendToCache/1024-shards-12    3.49µs ± 4%    2.06µs ± 4%  -40.86%  (p=0.000 n=100+93)
AppendToCache/8192-shards-12    4.04µs ± 3%    2.85µs ± 3%  -29.42%  (p=0.000 n=96+97)

name                          old alloc/op   new alloc/op   delta
AppendToCache/1-shards-12       28.1kB ± 0%    14.1kB ± 0%  -49.84%  (p=0.000 n=100+100)
AppendToCache/512-shards-12     28.2kB ± 1%    14.1kB ± 0%  -49.92%  (p=0.000 n=93+95)
AppendToCache/1024-shards-12    29.4kB ± 4%    14.4kB ± 4%  -50.96%  (p=0.000 n=100+94)
AppendToCache/8192-shards-12    34.0kB ± 2%    19.4kB ± 3%  -43.07%  (p=0.000 n=99+98)

name                          old allocs/op  new allocs/op  delta
AppendToCache/1-shards-12         21.0 ± 0%       2.0 ± 0%  -90.48%  (p=0.000 n=100+100)
AppendToCache/512-shards-12       22.0 ± 0%       3.0 ± 0%  -86.36%  (p=0.000 n=100+100)
AppendToCache/1024-shards-12      22.0 ± 0%       3.0 ± 0%  -86.36%  (p=0.000 n=100+100)
AppendToCache/8192-shards-12      22.0 ± 0%       3.0 ± 0%  -86.36%  (p=0.000 n=100+100)

```